Introduce hotkeystate-less TextInputV2
